Real ist för Integer

+++ KrustyKookies.sql
@@ -43,7 +43,7 @@ CREATE TABLE Recipes (
 CREATE TABLE Ingredients (
   material_name    TEXT,
   recipe_name      TEXT,
-  quantity         INTEGER CHECK (quantity > 0),
+  quantity         REAL CHECK (quantity > 0),
   PRIMARY KEY (material_name, recipe_name),
   FOREIGN KEY (material_name) REFERENCES RawMaterials(material_name),
   FOREIGN KEY (recipe_name) REFERENCES Recipes(recipe_name)
